workers layer and FXML added with the cards
