--- src/main/resources/dk/easv/views/worker-users-view.fxml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import io.github.palexdev.materialfx.controls.MFXButton?>
-<?import io.github.palexdev.materialfx.controls.MFXScrollPane?>
-<?import javafx.scene.Group?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.RowConstraints?>
-
-<GridPane prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="dk.easv.gui.controllers.WorkerViewController" stylesheets="@../css/global-style.css" styleClass="main_pane">
-   <columnConstraints>
-      <ColumnConstraints halignment="CENTER" hgrow="SOMETIMES" minWidth="10.0" />
-   </columnConstraints>
-   <rowConstraints>
-      <RowConstraints minHeight="10.0" valignment="CENTER" vgrow="SOMETIMES" />
-   </rowConstraints>
-   <children>
-      <Group GridPane.halignment="CENTER" GridPane.hgrow="NEVER" GridPane.vgrow="NEVER">
-         <children>
-            <GridPane prefHeight="400.0" prefWidth="800.0">
-              <columnConstraints>
-                <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-              </columnConstraints>
-              <rowConstraints>
-                <RowConstraints maxHeight="372.8000122070313" minHeight="10.0" prefHeight="372.8000122070313" vgrow="SOMETIMES" />
-                  <RowConstraints maxHeight="192.60001220703126" minHeight="10.0" prefHeight="26.399987792968773" vgrow="SOMETIMES" />
-              </rowConstraints>
-               <children>
-                  <MFXScrollPane fx:id="workerScrollPane" prefHeight="340.0" GridPane.hgrow="NEVER" GridPane.vgrow="NEVER">
-                   <content>
-                       <HBox fx:id="workers" prefHeight="99.0" prefWidth="575.0" />
-                   </content>
-                  </MFXScrollPane>
-                  <GridPane GridPane.rowIndex="1">
-                    <columnConstraints>
-                        <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                        <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                      <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                      <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                    </columnConstraints>
-                    <rowConstraints>
-                      <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-                    </rowConstraints>
-                     <children>
-                        <HBox prefHeight="100.0" prefWidth="200.0" GridPane.columnIndex="1">
-                           <children>
-                              <MFXButton fx:id="delete" prefHeight="26.0" prefWidth="231.0" text="Delete" />
-                           </children>
-                        </HBox>
-                        <HBox prefHeight="100.0" prefWidth="200.0" GridPane.columnIndex="2" />
-                        <HBox prefHeight="100.0" prefWidth="200.0" GridPane.columnIndex="3">
-                           <children>
-                              <MFXButton onAction="#goBack" prefHeight="26.0" prefWidth="215.0" text="Back" />
-                           </children>
-                        </HBox>
-                        <MFXButton fx:id="edit" prefHeight="26.0" prefWidth="274.0" text="Edit" />
-                     </children>
-                  </GridPane>
-               </children>
-            </GridPane>
-         </children>
-      </Group>
-   </children>
-</GridPane>
