+++ src/main/resources/dk/easv/views/workers-view.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<GridPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="700.0" prefWidth="1000.0" stylesheets="@../css/global-style.css" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="dk.easv.gui.controllers.BusinessViewController">
+<GridPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="700.0" prefWidth="1000.0" stylesheets="@../css/global-style.css" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="dk.easv.gui.controllers.WorkersViewController">
    <columnConstraints>
       <ColumnConstraints hgrow="SOMETIMES" maxWidth="445.199951171875" minWidth="10.0" prefWidth="100.0" />
       <ColumnConstraints hgrow="SOMETIMES" maxWidth="888.4000186920166" minWidth="10.0" prefWidth="825.6000793457031" />
@@ -34,17 +34,17 @@
          <children>
             <VBox fx:id="iconsVbox" alignment="TOP_CENTER" prefHeight="200.0" prefWidth="100.0">
                <children>
-                  <MFXButton fx:id="HomeLayer" onAction="#displayHome" styleClass="documentTest" text="Home">
+                  <MFXButton fx:id="HomeLayer" onAction="#displayHome" text="Home">
                      <VBox.margin>
                         <Insets bottom="10.0" top="10.0" />
                      </VBox.margin>
                   </MFXButton>
-                  <MFXButton fx:id="businessLayer" text="Business">
+                  <MFXButton fx:id="businessLayer" onAction="#displayBusiness" text="Business">
                      <VBox.margin>
                         <Insets bottom="10.0" top="10.0" />
                      </VBox.margin>
                   </MFXButton>
-                  <MFXButton fx:id="workersLayer" onAction="#displayWorkers" text="Workers">
+                  <MFXButton fx:id="workersLayer" text="Workers">
                      <VBox.margin>
                         <Insets bottom="10.0" top="10.0" />
                      </VBox.margin>
@@ -84,8 +84,8 @@
                   <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
                </columnConstraints>
                <rowConstraints>
-                  <RowConstraints maxHeight="293.0000061035156" minHeight="10.0" prefHeight="293.0000061035156" vgrow="SOMETIMES" />
-                  <RowConstraints maxHeight="262.6000122070312" minHeight="10.0" prefHeight="241.79999389648435" vgrow="SOMETIMES" />
+                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                  <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
                </rowConstraints>
                <children>
                   <GridPane GridPane.rowIndex="1">
@@ -93,12 +93,12 @@
                         <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
                      </columnConstraints>
                      <rowConstraints>
-                        <RowConstraints maxHeight="76.80001220703124" minHeight="10.0" prefHeight="45.60000381469726" vgrow="SOMETIMES" />
-                        <RowConstraints maxHeight="194.39999618530274" minHeight="10.0" prefHeight="194.39999618530274" vgrow="SOMETIMES" />
+                        <RowConstraints maxHeight="76.80001220703124" minHeight="10.0" prefHeight="12.000021553039545" vgrow="SOMETIMES" />
+                        <RowConstraints maxHeight="227.99997844696045" minHeight="10.0" prefHeight="227.99997844696045" vgrow="SOMETIMES" />
                      </rowConstraints>
                      <children>
-                        <Label text="Documents" />
-                        <MFXTableView fx:id="documentsTable" footerVisible="false" prefHeight="198.0" prefWidth="830.0" GridPane.rowIndex="1" />
+                        <MFXTableView fx:id="projectTable" footerVisible="false" prefHeight="230.0" prefWidth="830.0" GridPane.rowIndex="1" />
+                        <Label fx:id="projectLabel" text="Project" />
                      </children>
                   </GridPane>
                   <GridPane>
@@ -106,40 +106,25 @@
                       <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
                     </columnConstraints>
                     <rowConstraints>
-                      <RowConstraints maxHeight="267.1999877929687" minHeight="10.0" prefHeight="267.1999877929687" vgrow="SOMETIMES" />
-                      <RowConstraints maxHeight="141.60001220703126" minHeight="10.0" prefHeight="25.600012207031284" vgrow="NEVER" />
+                      <RowConstraints maxHeight="239.19998779296876" minHeight="10.0" prefHeight="239.19998779296876" vgrow="SOMETIMES" />
+                      <RowConstraints maxHeight="129.60001220703126" minHeight="10.0" prefHeight="29.600012207031256" vgrow="SOMETIMES" />
                     </rowConstraints>
                      <children>
-                        <MFXScrollPane fx:id="customersScrollPane">
+                        <MFXScrollPane fx:id="workerScrollPane">
                            <content>
-                              <HBox fx:id="Customers" prefHeight="100.0" prefWidth="200.0" />
-                           </content>
-                        </MFXScrollPane>
+                              <HBox fx:id="workers" prefHeight="100.0" prefWidth="200.0" />
+                           </content></MFXScrollPane>
                         <GridPane GridPane.rowIndex="1">
                           <columnConstraints>
                             <ColumnConstraints halignment="CENTER" hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
                               <ColumnConstraints halignment="CENTER" hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                              <ColumnConstraints halignment="CENTER" hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
                           </columnConstraints>
                           <rowConstraints>
-                            <RowConstraints minHeight="10.0" prefHeight="30.0" valignment="CENTER" vgrow="SOMETIMES" />
+                            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
                           </rowConstraints>
                            <children>
-                              <MFXButton fx:id="createCustomer" onAction="#createCustomer" text="New Customer">
-                                 <GridPane.margin>
-                                    <Insets top="5.0" />
-                                 </GridPane.margin>
-                              </MFXButton>
-                              <MFXButton fx:id="edit" text="Edit" GridPane.columnIndex="1">
-                                 <GridPane.margin>
-                                    <Insets top="5.0" />
-                                 </GridPane.margin>
-                              </MFXButton>
-                              <MFXButton fx:id="delete" text="Delete" GridPane.columnIndex="2">
-                                 <GridPane.margin>
-                                    <Insets top="5.0" />
-                                 </GridPane.margin>
-                              </MFXButton>
+                              <MFXButton text="Edit" />
+                              <MFXButton text="Delete" GridPane.columnIndex="1" />
                            </children>
                         </GridPane>
                      </children>
